extract turn based render to hook
