+++ src/app/page.tsx
@@ -1,24 +1,13 @@
 'use client'
 import { wave } from '@/animations';
-import { Heading, VStack, HStack, Link, List, ListItem, ListIcon, Button, Text, Stack, useBreakpoint, Box, useDimensions, Slide, SlideFade } from '@chakra-ui/react';
-import { EmailIcon, AtSignIcon } from '@chakra-ui/icons';
+import { Heading, VStack, HStack, useBreakpoint, Box, SlideFade } from '@chakra-ui/react';
 import Image from 'next/image'
 import NavItemsList, { NavItem } from '@/components/NavItemList';
 import { FaGithub, FaLinkedin } from 'react-icons/fa';
-import TypeText from '@/components/TypeText';
-import { memo, useEffect, useMemo, useRef, useState } from 'react';
+import { useMemo, useState } from 'react';
 import TypeParagraph from '@/components/TypeParagraph';
+import { useTurnBasedRender } from '@/hooks/useTurnBasedRender';
 
-const LetsConnect = () => {
-  return (
-    <>
-      <Heading size="lg" mb={2}>🤝 Let's Connect</Heading>
-      <Text>
-        Got an idea to share or just want to say hello? Shoot me an <Link href="mailto:mlpierce22@gmail.com" color="teal.500"><EmailIcon mx="2px" /> email</Link> or let's connect on <Link href="https://linkedin.com/in/malapier" isExternal color="teal.500"><AtSignIcon mx="2px" /> LinkedIn</Link>.
-      </Text>
-    </>
-  )
-}
 const HomePage = () => {
 
   const breakpoint = useBreakpoint();
@@ -32,21 +21,6 @@ const HomePage = () => {
     },
     { id: "connect-linkedin-button", name: 'Connect on LinkedIn', icon: <FaLinkedin />, path: "https://linkedin.com/in/malapier", variant: 'solid', isExternal: true }
   ]
-
-  // const quickFacts = [
-  //   { icon: '•', desc: 'Based in San Francisco, CA' },
-  //   { icon: '•', desc: 'Former student athlete at UC Santa Cruz' },
-  //   { icon: '•', desc: 'Full-stack software engineer with a focus on user experience' },
-  //   { icon: '•', desc: 'Experience building ML applications for enterprise' }
-  // ];
-
-  // const professionalInterests = [
-  //   { icon: '•', desc: 'Innovating with edge technologies' },
-  //   { icon: '•', desc: 'Crafting user-centric solutions' },
-  //   { icon: '•', desc: 'Contributing to open-source projects' },
-  //   { icon: '•', desc: 'Exploring the potential of LLMs' }
-  // ];
-
 
   const header = useMemo(() => (
     "Hey there, I'm Mason! 👋"
@@ -62,15 +36,7 @@ const HomePage = () => {
 
   const typeSpeedSeconds = 1;
   const pageLoadDelaySeconds = 0.5;
-  const [turnCounter, setTurnCounter] = useState(0);
-
-  const setNextTurn = (nextTurn: number) => {
-    setTurnCounter(nextTurn);
-  }
-
-  const isMyTurn = (myPlace: number) => {
-    return turnCounter >= myPlace;
-  }
+  const { setNextTurn, isMyTurn } = useTurnBasedRender();
 
   return (
     <VStack p={10} spacing={10}>
@@ -87,7 +53,7 @@ const HomePage = () => {
       <VStack spacing={10}>
         {isMyTurn(0) && (
           <TypeParagraph paragraph={header} typeSpeedSeconds={typeSpeedSeconds} delaySeconds={pageLoadDelaySeconds} doneTypingCallback={() => setNextTurn(1)} align='center'>
-            <Heading size={{ base: "lg", md: "3xl" }}>
+            <Heading size={{ base: "lg", md: "3xl" }} className='leading-snug'>
               {header}
             </Heading>
           </TypeParagraph>
